requirements.txt
clarifai==2.6.2
google-cloud-speech==1.2.0
kthread==0.2.2
lxml==4.3.3
Pillow==2.3.0
protobuf==3.7.1
pymongo==3.7.2
selenium==3.141.0
urllib3==1.25.2
